+++ src/set.rs
@@ -9,18 +9,33 @@ pub trait SetOps<T>: Bounded<T> {
 
     /// Compute the intersection of `self` and `other`.
     ///
+    /// # Notes
+    /// If the `self` is degenerate, just return it as is.
+    ///
     /// # Errors
-    /// Return a pair of original values if at least one of [`IntoBounds::into_bounds`] fails.
-    fn intersect<R>(self, other: R) -> Result<Self, (Self, R)>
+    /// Return `other` if it is degenerate (for the normal `self`).
+    fn intersect<R>(self, other: R) -> Result<Self, R>
     where
         T: Ord,
         Self: From<Self::Error>,
         R: IntoBounds<T> + From<R::Error>,
     {
-        let ((self_start, self_end), (other_start, other_end)) = pair_into_bounds(self, other)?;
+        let ((a, b), (c, d)) = match pair_into_bounds(self, other) {
+            Ok(bounds) => bounds,
+            Err((left, right)) => {
+                let left_degenerate = left.into_bounds().err();
+                // if _left_ is degenerate, return it as is
+                return left_degenerate.map(Self::from).ok_or({
+                    // otherwise, _left_ is valid and _right_ is degenerate,
+                    // but we cannot construct a value of `Self`
+                    // from a degenerate _right_, so just return it as `Err`
+                    right
+                });
+            }
+        };
 
-        let start = self_start.max(other_start);
-        let end = self_end.min(other_end);
+        let start = a.max(c);
+        let end = b.min(d);
 
         Ok(Self::from_bounds((start, end)))
     }
@@ -31,18 +46,27 @@ pub trait SetOps<T>: Bounded<T> {
     /// Otherwise (when the intervals are disjoint),
     /// return a [pair][OneOrPair::Pair] of pairs of ordered ranges
     ///
-    /// # Errors
-    /// Return a pair of original values if at least one of [`IntoBounds::into_bounds`] fails.
-    fn union<R>(self, other: R) -> Result<OneOrPair<Self>, (Self, R)>
+    /// # Notes
+    /// If at least one of [`IntoBounds::into_bounds`] fails (the value is degenerate),
+    /// return the other value wrapped in [`OneOrPair::One`].
+    fn union<R>(self, other: R) -> OneOrPair<Self>
     where
         T: Ord,
         Self: From<Self::Error>,
         R: IntoBounds<T> + From<R::Error>,
     {
-        let ((self_start, self_end), (other_start, other_end)) = pair_into_bounds(self, other)?;
+        let ((a, b), (c, d)) = match pair_into_bounds(self, other) {
+            Ok(bounds) => bounds,
+            Err((left, right)) => {
+                // - if _right_ is degenerate, return _left_ as is (either it is degenerate or not)
+                // - if _right_ is not degenerate (then _left_ should be degenerate, since the pair failed),
+                //   then transform the _right_ into `Self` and return it.
+                return OneOrPair::One(right.into_bounds().map_or(left, Self::from_bounds));
+            }
+        };
 
-        let [min_start, max_start] = minmax(self_start, other_start);
-        let [min_end, max_end] = minmax(self_end, other_end);
+        let [min_start, max_start] = minmax(a, c);
+        let [min_end, max_end] = minmax(b, d);
 
         let are_disjoint = {
             let (l, r) = (max_start.as_ref(), min_end.as_ref());
@@ -54,32 +78,42 @@ pub trait SetOps<T>: Bounded<T> {
             empty_intersection && !empty_gap
         };
 
-        let one_or_pair = if are_disjoint {
+        if are_disjoint {
             OneOrPair::Pair((
                 Self::from_bounds((min_start, min_end)),
                 Self::from_bounds((max_start, max_end)),
             ))
         } else {
             OneOrPair::One(Self::from_bounds((min_start, max_end)))
-        };
-        Ok(one_or_pair)
+        }
     }
 
-    /// The smallest span containing both `self` and `other`.
+    #[must_use = "Computing the enclosure returns a new value from the inputs"]
+    /// The smallest span containing both `self` and `other`,
+    /// also covering possible 'gap' between them.
     ///
-    /// # Errors
-    /// Return a pair of original values if at least one of [`IntoBounds::into_bounds`] fails.
-    fn enclosure<R>(self, other: R) -> Result<Self, (Self, R)>
+    /// # Notes
+    /// If at least one of [`IntoBounds::into_bounds`] fails (the value is degenerate),
+    /// return the other value wrapped in [`OneOrPair::One`].
+    fn enclosure<R>(self, other: R) -> Self
     where
         T: Ord,
         Self: From<Self::Error>,
         R: IntoBounds<T> + From<R::Error>,
     {
-        let ((self_start, self_end), (other_start, other_end)) = pair_into_bounds(self, other)?;
+        let ((a, b), (c, d)) = match pair_into_bounds(self, other) {
+            Ok(bounds) => bounds,
+            Err((left, right)) => {
+                // - if _right_ is degenerate, return _left_ as is (either it is degenerate or not)
+                // - if _right_ is not degenerate (then _left_ should be degenerate, since the pair failed),
+                //   then transform the _right_ into `Self` and return it.
+                return right.into_bounds().map_or(left, Self::from_bounds);
+            }
+        };
 
-        let start = self_start.min(other_start);
-        let end = self_end.max(other_end);
-        Ok(Self::from_bounds((start, end)))
+        let start = a.min(c);
+        let end = b.max(d);
+        Self::from_bounds((start, end))
     }
 
     // TODO: contains
@@ -99,7 +133,7 @@ pub trait SetOps<T>: Bounded<T> {
         &'a Self: IntoBounds<&'a T>,
         R: IntoBounds<&'a T>,
     {
-        // if at least one of the intervals is empty, they cannot be disjoint
+        // if at least one of the intervals is degenerate, they cannot be disjoint
 
         let Ok((self_start, self_end)) = self.into_bounds() else {
             return false;
@@ -207,10 +241,8 @@ mod tests {
 
         let a = interval!(U: i32);
         let b = interval!(0: i32);
-        assert!(matches!(
-            a.intersect(b).unwrap_err(),
-            (interval!(U), interval!(0))
-        ));
+        assert!(matches!(a.intersect(b).unwrap_err(), interval!(0)));
+        assert!(matches!(b.intersect(a).unwrap(), interval!(0)));
     }
 
     #[test]
@@ -232,7 +264,8 @@ mod tests {
 
         let a = interval!((2, =4));
         let b = interval!((=3, 1));
-        assert_eq!(a.intersect(b).unwrap_err(), (a, b));
+        assert_eq!(a.intersect(b).unwrap_err(), b);
+        assert_eq!(b.intersect(a).unwrap(), b);
     }
 
     #[test]
@@ -252,29 +285,26 @@ mod tests {
     fn enclosure() {
         let a = interval!([3, 7]);
         let b = interval!((5, 10));
-        assert_eq!(a.enclosure(b).unwrap(), interval!((=3, 10)));
+        assert_eq!(a.enclosure(b), interval!((=3, 10)));
 
         let a = interval!(<5);
         let b = interval!(>3);
-        assert_eq!(a.enclosure(b).unwrap(), Interval::Full);
+        assert_eq!(a.enclosure(b), Interval::Full);
 
         let a = interval!(<=-100);
         let b = interval!(>=100);
-        assert_eq!(a.enclosure(b).unwrap(), Interval::Full);
+        assert_eq!(a.enclosure(b), Interval::Full);
 
         let a = interval!([1, 2]);
         let b = interval!([3, 4]);
-        assert_eq!(a.enclosure(b).unwrap(), interval!([1, 4]));
+        assert_eq!(a.enclosure(b), interval!([1, 4]));
     }
 
     #[test]
     fn union_touching() {
         let a = interval!([1, 2]);
         let b = interval!([2, 4]);
-        assert_eq!(
-            a.union(b).unwrap().into_single().unwrap(),
-            interval!([1, 4])
-        );
+        assert_eq!(a.union(b).into_single().unwrap(), interval!([1, 4]));
 
         let a = interval!((1, 2));
         let b = interval!([2, 4]);
@@ -282,10 +312,7 @@ mod tests {
 
         let a = interval!([1, 2]);
         let b = interval!((2, 4));
-        assert_eq!(
-            a.union(b).unwrap().into_single().unwrap(),
-            interval!((=1, 4))
-        );
+        assert_eq!(a.union(b).into_single().unwrap(), interval!((=1, 4)));
 
         let a = interval!((1, 2));
         let b = interval!((2, 4));